Change the front end code to use Angular
To give the application some more structure, I rewrote the entire front
end to use AngularJS.
I've also added in the `ng-infinite-scroll` module which makes
lazyLoading even easier.

Right now with this commit, the settings modal is removed. This will be
added in soon, I just wanted to get this base pushed up.