Fix a bug with the settings
I forgot to not the `dontOpenBrowserOnStart` value, and I passed the
params to update settings incorrectly.
I also removed the timer from fetching peers. Right now it will only
fetch on page load, but I'll soon implement web sockets so that the go
back end will send a message when peers are updated.